chore: keep only homepage in the homepage.dart
