+++ lib/src/screens/homepage.dart
@@ -1,10 +1,4 @@
 import 'package:flutter/material.dart';
-import 'dart:convert';
-
-Future getSpecieData(String name, BuildContext context) async {
-  String data = await DefaultAssetBundle.of(context).loadString("data/"+name+".json");
-  return json.decode(data);
-}
 
 final List<String> animals = ['fish', 'reptiles', 'mammals', 'birds', 'insets', 'humans'];
 final String title = "Evolutionary";
@@ -12,17 +6,6 @@ final ThemeData theme = ThemeData(
   primaryColor: Color(0xFF118855),
   accentColor: Colors.yellow,
 );
-
-class Evolutionary extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(title: title, theme: theme, home: Scaffold(
-      appBar: AppBar(title: Text(title), elevation: 0,),
-      body: HomePage()
-    ));
-  }
-}
-
 
 class HomePage extends StatelessWidget {
   @override
